Fix bugs around metavariable conflict resolution
This eliminate some spurious conflicts that should not be generated by
also fixes crashes on examples that created cycles in substitutions.

Wait until substitution step for solving conflicts.
Do not panic on cycles in substitution, instead mark metavariable as
conflicting for ins_subst, and behave as identity for del_subst.
We also wait until the end of the substitution to remove potential metavariable
conflicts from deletion trees. This allows to never eliminate a
replacement that was finally never inserted anywhere.