Change exit code on panics to -3
To follow the convention of `git merge-file`, we want the error codes
between 0 and 127 to represent the number of conflicts that are left to
resolve.

However, by default Rust returns 101 on panic, so we insert a custom
panic hook to return -3 instead, and optimize out the unwinding step that we
did not use anyway.